+++ src/v/kafka/protocol/legacy_message.h
@@ -13,7 +13,7 @@
 #include "bytes/iobuf.h"
 #include "bytes/iobuf_parser.h"
 #include "hashing/crc32.h"
-#include "kafka/server/logger.h"
+#include "kafka/protocol/logger.h"
 #include "model/compression.h"
 #include "model/fundamental.h"
 #include "model/timestamp.h"
