+++ src/v/kafka/protocol/logger.cc
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "kafka/server/logger.h"
+#include "kafka/protocol/logger.h"
 
 namespace kafka {
 ss::logger klog("kafka");
